requirements.txt
absl-py==2.3.1
altgraph==0.17.4
anitya_schema==2.3.0
asgiref==3.11.0
astunparse==1.6.3
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
blinker==1.9.0
cbor2==5.8.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
crispy-bootstrap5==2025.6
crochet==2.1.1
cryptography==46.0.3
daphne==4.2.1
decorator==5.2.1
defusedxml==0.7.1
dj-database-url==3.1.0
Django==4.2.8
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-extensions==4.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
fedora_messaging==3.9.0
fedora_messaging_the_new_hotness_schema==1.3.0
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
gssapi==1.10.1
gunicorn==24.1.1
h5py==3.15.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.11
ImageIO==2.37.2
Incremental==24.11.0
jax==0.8.2
jaxlib==0.8.2
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.13.0
koji==1.35.3
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.46.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
namex==0.1.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.6
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pefile==2023.2.7
pika==1.3.2
pillow==11.0.0
platformdirs==4.5.1
pooch==1.8.2
protobuf==6.33.2
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.10.1
PyMatting==1.1.14
pyOpenSSL==25.3.0
pyreadline3==3.5.4
python-bugzilla==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
pywin32-ctypes==0.2.3
redis==5.3.1
referencing==0.37.0
rembg==2.0.69
requests==2.32.5
requests-gssapi==1.4.0
rich==14.2.0
rpds-py==0.30.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
the-new-hotness==1.3.0
threadpoolctl==3.6.0
tifffile==2025.12.20
tomli==2.3.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.2
Werkzeug==3.1.4
wheel==0.45.1
whitenoise==6.11.0
wrapt==2.0.1
zope.interface==8.2
